JAMES-2186 Correct JMAP integration test
The test was incorrect, as it looked for the messageId of the message in the Outbox, while the INBOX messageId is different.

The meaning is, if the test was fast enough to pass before message is removed from outbox, it pass. If message is remove from outbox before the test execute GetMessageList, it fails.

I corrected the scenario of this test.