JAMES-2257 OldKeyword as Keywords
We should limit at the maximum the spread of the knowledge of OldKeywords.

In some situation an OldKeyword is a relative view of the data (during updates), sometine an absolute view of the data (while creating it).
When we create a message we can simply concider the OldKeywords as a normal Keywords.

See later on, it brings significant simplifications...