JAMES-2408 Allow one to await the execution of ElasticSearchIndexer::deleteAll
Note that this is used in case of mailbox deletions to delete messages that are anyway not accessible anymore, as a background task.

However, we might have an embedded ElasticSearch not being active while the deleteByQuery is not finished yet, leading to test failures. Awaiting the task solves this issue.