JAMES-2641 Make StoreMailboxManager::batchSizes final
To do this we need to let the injection system providing the built value.

Note that we inject a more generic "MailboxManagerConfiguration" class in order to make
future parameters addition easier in the future.

Note that Sprint (& memory) did not configure the batch size. Thus to reduce diff size the constructors directly provide default value as a temporary solution.