MAILBOX-374 Turn off exclusive - auto-deletes queues
In the face of network errors, we can not achieve a 'safety' and 'liveness' property.

 - 'liveness': A James node that is not failmed always have its Key registration queue
 - 'safety': The "key registration queue" gets deleted when there is not more consumer.

If the network partition exceeded the configured heartbeat, then the queue is going to be deleted and James is going to fail (as its queue disapeared).

We choose the safer option of not having RabbitMQ responsible of automatic queue deletion. The admin is going to be responsible of queue cleanup after crash.