JAMES-2794 Avoid calling *block* in doOnNext
Reactor is complaining that block was called in a parallel thread around 4-8%
of the time. Switching the subscription scheduler did not modify this behaviour.

However, we do not need to actually wait the browse update to be performed
before returning: this can perfectly be a background process.

This actually solves stability issues of browseAndDequeueShouldCombineWellWhenDifferentSlices