JAMES-2856 Fix QuotaRoot serialization
Mailbox event serialization relies on their own logic for quotaRoot deserialization,
which is duplicated from the standard one, and buggy (no management for domain)

We need to have mailbox event serialization reusing logic for quotaRoot deserialization.

That being said, the QuotaRootResolver, encapsulating this logic, comes with too much dependencies,
and thus is not convenient for testing. That's why we split the deserializer away from the resolver
(no dependencies)

Feature relying on this is limited to quotaSearch domainFiltering on top of distributed James product.