JAMES-2080 Address headers do not require NESTED ElasticSearch mapping
Nested objects are indexed as separate documents which incurs an overhead
at index time and also has a cost upon search as ElasticSearch needs to
perform joins between the nested documents and the original one.

It turns out that the extra coherence between mailbox name and address
brought by nested objects is not needed functionally speaking and thus we
can get rid of it easily.

Note that this change needs a re-indexing of previously existing data