JAMES-2927 Avoid un-intentional tumbstones in Cassandra mailqueue view projections
A tumbstone is created when a null value is specified in a prepared statement.

This is due to the fact that null has the meaning `remove` and not the meaning `unspecified`, which is represented by no binding at all.

Of course unwanted tumbstones occurs with a performance cost.

The recommended method for fixing this is to separate nullable parts of the statements in other prepared statements, that can logically grouped together without overhead using batch statements.

Read this for further information: https://thelastpickle.com/blog/2016/09/15/Null-bindings-on-prepared-statements-and-undesired-tombstone-creation.html