JAMES-2870 Remove ZooKeeper Uid & ModSeq generators
This module allows to generate sequence numbers for things like IMAP
modseq or uid. It has been implemented to fulfil the lack of support
for this kind of feature in HBase.

Cassandra has the same kind of problems, as most distributed
database, but we solved the problem without requiring yet another
complex middleware.

As HBase is no longer part of the codebase and to my knowledge the
component is not used anywhere else, I propose to mark it as
deprecated for 3.4 in order to target a removal for 3.5