JAMES-2983 Add MailboxId as part of MailboxResponse
Currently, several webAdmin APIs relies on mailboxId in their endpoints.

That is a good choice, as mailboxId is immutable and thus more predictable than mailboxPath.

However, mailboxId is not directly exposed to the end user, who can get confused.

Currently, using webadmin, I have no way of linking mailboxPath and mailboxId

Proposal is to add mailboxId as part of the MailboxResponse object.