JAMES-3102 Allow to bypass alias user existence check
User A had been changing name (to user B) and both entries exist in the LDAP.

We want User B to send and receive mails from UserA address.

We are using LDAP repository. We cannot delete user A from the LDAP. We can not create the alias as user A exists.

Thus we want to be able to 'force' the alias insertion despite the rule that a user can't be used as an alias source.

Traffic of user A will effectively be redirected to user B. Moreover User B will still be able to send mails using user A identify (not possible with forwards).

The proposal is to introduce a new query parameter allowing to bypass the check, while keeping the behaviour untouched otherwise.

Example:

curl -XPUT http://ip:port/address/aliases/user@domain.com/sources/alias@domain.com?bypassUserCheck