JAMES-3105 Add a service for recomputing mailbox counters
Algorithm:

 - List existing mailboxes
 - List their messages
 - Check them against their source of truth
 - Compute mailbox counter values
 - And reset the value of the counter if needed

Limitations:

This won't be friendly in the face of concurrent operations, which
would be ignored during the recomputation of the counter of the
given mailbox. However the source of truth is unaffected hence,
upon rerunning the algorithm, the result will be eventually correct.

We rely on the "listing messages by mailbox" projection (that we
recheck). Missing entries in there will be ignored until the given
projection is healed (currently unsupported).